feat: move http-exception out of utils
