+++ deno_dist/middleware/basic-auth/index.ts
@@ -1,8 +1,8 @@
+import { HTTPException } from '../../http-exception.ts'
 import type { HonoRequest } from '../../request.ts'
 import type { MiddlewareHandler } from '../../types.ts'
 import { timingSafeEqual } from '../../utils/buffer.ts'
 import { decodeBase64 } from '../../utils/encode.ts'
-import { HTTPException } from '../../utils/http-exception.ts'
 
 const CREDENTIALS_REGEXP = /^ *(?:[Bb][Aa][Ss][Ii][Cc]) +([A-Za-z0-9._~+/-]+=*) *$/
 const USER_PASS_REGEXP = /^([^:]*):(.*)$/
