+++ deno_dist/middleware/jwt/index.ts
@@ -1,5 +1,5 @@
+import { HTTPException } from '../../http-exception.ts'
 import type { MiddlewareHandler } from '../../types.ts'
-import { HTTPException } from '../../utils/http-exception.ts'
 import { Jwt } from '../../utils/jwt/index.ts'
 import type { AlgorithmTypes } from '../../utils/jwt/types.ts'
